+++ usecase/usecase.go
@@ -156,100 +156,7 @@ func (uc *UserSignUp) SaveUserProfile(login string) (int, error) {
 		Interests:   interests,
 	}
 
-	fmt.Println(fmt.Errorf("Profile: %+v", profile))
+	fmt.Println(fmt.Errorf("profile: %+v", profile))
 
 	return uc.userRepo.StoreProfile(profile)
-}
-
-type UserCheckSession struct {
-	sessionRepo repository.SessionRepository
-}
-
-func NewUserCheckSessionUseCase(sessionRepo repository.SessionRepository) *UserCheckSession {
-	return &UserCheckSession{sessionRepo: sessionRepo}
-}
-
-func (uc *UserCheckSession) CheckSession(sessionId string) (int, error) {
-	userIdStr, err := uc.sessionRepo.GetSession(sessionId)
-	if err != nil {
-		return -1, err
-	}
-
-	userId, err := strconv.Atoi(userIdStr)
-	if err != nil {
-		return -1, model.ErrInvalidSessionId
-	}
-	return userId, nil
-}
-
-type UserLogOut struct {
-	userRepo    repository.UserRepository
-	sessionRepo repository.SessionRepository
-}
-
-func NewUserLogOutUseCase(userRepo repository.UserRepository, sessionRepo repository.SessionRepository) *UserLogOut {
-	return &UserLogOut{
-		userRepo:    userRepo,
-		sessionRepo: sessionRepo,
-	}
-}
-
-func (ul *UserLogOut) Logout(sessionId string) error {
-	userIdStr, err := ul.sessionRepo.GetSession(sessionId)
-	if err != nil {
-		return err
-	}
-
-	err = ul.sessionRepo.DeleteSession(sessionId)
-	if err != nil {
-		return model.ErrDeleteSession
-	}
-
-	userId, err := strconv.Atoi(userIdStr)
-	if err != nil {
-		return model.ErrInvalidSessionId
-	}
-
-	err = ul.userRepo.DeleteSession(userId)
-	if err != nil {
-		return model.ErrDeleteSession
-	}
-
-	return nil
-}
-
-type UserDelete struct {
-	userRepo repository.UserRepository
-}
-
-func NewUserDeleteUseCase(userRepo repository.UserRepository) *UserDelete {
-	return &UserDelete{userRepo: userRepo}
-}
-
-func (ud *UserDelete) DeleteUser(userId int) error {
-	return ud.userRepo.DeleteUserById(userId)
-}
-
-type GetProfile struct {
-	userRepo repository.UserRepository
-}
-
-func NewGetProfileUseCase(userRepo repository.UserRepository) *GetProfile {
-	return &GetProfile{userRepo: userRepo}
-}
-
-func (gp *GetProfile) GetProfile(userId int) (model.Profile, error) {
-	return gp.userRepo.GetProfileById(userId)
-}
-
-type GetProfilesForUser struct {
-	userRepo repository.UserRepository
-}
-
-func NewGetProfilesForUserUseCase(userRepo repository.UserRepository) *GetProfilesForUser {
-	return &GetProfilesForUser{userRepo: userRepo}
-}
-
-func (gp *GetProfilesForUser) GetProfiles(forUserId int) ([]model.Profile, error) {
-	return gp.userRepo.GetProfilesByUserId(forUserId)
 }
